+++ src/main/CharacterSets.ts
@@ -12,12 +12,14 @@ export class CharacterSet {
   NEGATIVE: string;
   POSITIVE: string;
   DOT: string;
+  TRADITIONAL_DOT: string;
+  TRADITIONAL_FLOATING_UNIT: string;
   CONTRACTION: string[];
   CONTRACTION_200: string;
   useTraditional: boolean;
   useCapital: boolean;
 
-  constructor(Base: string, Magnitudes: string, Sign: string, Dot: string, contraction: string, useTraditional: boolean, useCapital: boolean) {
+  constructor(Base: string, Magnitudes: string, Sign: string, Dot: string, traditionalFloatingUnit: string, contraction: string, useTraditional: boolean, useCapital: boolean) {
     this.BASE = Base;
     this.ZERO = Base[0];
     this.TEN = Magnitudes[0];
@@ -28,9 +30,9 @@ export class CharacterSet {
     this.TRILLION = Magnitudes[5];
     this.POSITIVE = Sign[0];
     this.NEGATIVE = Sign[1];
-    this.DOT = Dot;
-    this.useTraditional = useTraditional;
-    this.useCapital = useCapital;
+    this.DOT = Dot[0];
+    this.TRADITIONAL_DOT = Dot[1];
+    this.TRADITIONAL_FLOATING_UNIT = traditionalFloatingUnit;
     this.CONTRACTION = [];
     this.CONTRACTION[2] = contraction[0];
     this.CONTRACTION[3] = contraction[1];
@@ -40,17 +42,18 @@ export class CharacterSet {
     this.CONTRACTION[7] = contraction[5];
     this.CONTRACTION[8] = contraction[6];
     this.CONTRACTION[9] = contraction[7];
-
     this.CONTRACTION_200 = contraction[8];
+    this.useTraditional = useTraditional;
+    this.useCapital = useCapital;
   }
 }
 
 const allCharacterSets = [
   //
-  new CharacterSet('零一二三四五六七八九', '十百千万亿兆', '正负', '点', '廿卅卌圩圆进枯桦皕', false, false),
-  new CharacterSet('零壹贰叁肆伍陆柒捌玖', '拾佰仟萬億兆', '正负', '点', '念', false, true),
-  new CharacterSet('零一二三四五六七八九', '十百千萬億兆', '正負', '點', '廿卅卌圩圓進枯樺皕', true, false),
-  new CharacterSet('零壹貳參肆伍陸柒捌玖', '拾佰仟萬億兆', '正負', '點', '念', true, true),
+  new CharacterSet('零一二三四五六七八九', '十百千万亿兆', '正负', '点又', '分厘毫秒忽微纤沙尘埃渺漠', '廿卅卌圩圆进枯桦皕', false, false),
+  new CharacterSet('零壹贰叁肆伍陆柒捌玖', '拾佰仟萬億兆', '正负', '点又', '分厘毫秒忽微纤沙尘埃渺漠', '念', false, true),
+  new CharacterSet('零一二三四五六七八九', '十百千萬億兆', '正負', '點又', '分釐毫秒忽微纖沙塵埃渺漠', '廿卅卌圩圓進枯樺皕', true, false),
+  new CharacterSet('零壹貳參肆伍陸柒捌玖', '拾佰仟萬億兆', '正負', '點又', '分釐毫秒忽微纖沙塵埃渺漠', '念', true, true),
 ];
 
 export function getCharacterSet(useTraditional = false, useCapital = false): CharacterSet {